+++ packages/expo-router/src/layouts/JSStack.tsx
@@ -49,6 +49,7 @@ const JSStack = unstable_integrateWithRouter<
   object,
   StackNavigatorCreateProps
 >(unstable_createStandardStackNavigator, StackRouter, {
+  activityDefaultThreshold: 2,
   createProps: unstable_createPropsForJSStack,
 });