+++ packages/expo-router/src/layouts/TopTabsClient.tsx
@@ -35,6 +35,7 @@ const TopTabs = unstable_integrateWithRouter<
   TabRouterOptions,
   MaterialTopTabNavigatorCreateProps
 >(createStandardMaterialTopTabNavigator, TabRouter, {
+  activityDefaultThreshold: 1,
   processDescriptors: appendMissingPlaceholderTabDescriptors,
   processState: appendMissingPlaceholderTabRoutes,
   createProps: ({ state, dispatch, dispatchSync }) => ({